ticket_master_backend/requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
attrs==25.4.0
autopage==0.6.0
cachetools==5.5.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cliff==4.13.1
cmd2==2.5.11
colorama==0.4.6
cryptography==46.0.4
debtcollector==3.0.0
decorator==5.2.1
Django>=5.1.0
django-cors-headers==4.9.0
django-debug-toolbar==6.0.0
django-rest==0.8.7
djangorestframework>=3.15.0
djangorestframework_simplejwt==5.5.1
dogpile.cache==1.5.0
geopy==2.4.1
googleapis-common-protos==1.69.1
greenlet==3.3.1
gunicorn==25.0.2
idna==3.11
iniconfig==2.3.0
iso8601==2.1.0
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keystoneauth1==5.13.0
msgpack==1.1.2
mysqlclient==2.2.6
netaddr==1.3.0
openstacksdk==4.6.0
os-service-types==1.7.0
osc-lib==4.0.0
oslo.config==9.7.1
oslo.i18n==6.5.1
oslo.serialization==5.7.0
oslo.utils==8.2.0
packaging==26.0
pbr==7.0.3
pi==0.1.2
platformdirs==4.5.1
pluggy==1.6.0
prettytable==3.17.0
protobuf==5.29.3
psutil==7.2.2
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.14.0
pycparser==3.0
pydantic>=2.5.0
pydantic_core>=2.14.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.1
pyperclip==1.11.0
Pillow==10.2.0
qrcode==7.4.2
pytest==9.0.2
python-cinderclient==9.7.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-slugify==8.0.4
PyYAML==6.0.2
referencing==0.37.0
requests==2.32.3
requestsexceptions==1.4.0
rfc3986==2.0.0
rpds-py==0.30.0
setuptools==81.0.0
six==1.17.0
SQLAlchemy==2.0.45
sqlparse==0.5.3
stevedore==5.6.0
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2024.2
urllib3==2.6.3
waitress==3.0.2
wcwidth==0.6.0
wfastcgi==3.0.0
whitenoise==6.9.0
wrapt==2.1.1
